+++ src/table/table-role/__tests__/use-grid-navigation.test.tsx
@@ -1,130 +1,11 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-import React, { useRef, useState } from 'react';
+import React, { useRef } from 'react';
 import { render, fireEvent } from '@testing-library/react';
 import { useGridNavigation } from '../../../../lib/components/table/table-role';
 import { KeyCode } from '../../../../lib/components/internal/keycode';
-
-const mockObserver = {
-  observe: jest.fn(),
-  disconnect: jest.fn(),
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  callback: (mutationRecords: MutationRecord[]) => {},
-};
-const originalMutationObserver = global.MutationObserver;
-
-beforeEach(() => {
-  global.MutationObserver = function MutationObserver(callback: (mutationRecords: MutationRecord[]) => void) {
-    mockObserver.callback = callback;
-    return mockObserver;
-  } as any;
-});
-
-afterEach(() => {
-  global.MutationObserver = originalMutationObserver;
-});
-
-function TestTable<T extends object>({
-  tableRole = 'grid',
-  columns,
-  items,
-  startIndex = 0,
-  pageSize = 2,
-}: {
-  tableRole?: 'grid' | 'table';
-  columns: { header: React.ReactNode; cell: (item: T) => React.ReactNode }[];
-  items: T[];
-  startIndex?: number;
-  pageSize?: number;
-}) {
-  const tableRef = useRef<HTMLTableElement>(null);
-  useGridNavigation({
-    keyboardNavigation: tableRole === 'grid',
-    pageSize,
-    getTable: () => tableRef.current,
-  });
-  return (
-    <table role={tableRole} ref={tableRef}>
-      <thead>
-        <tr aria-rowindex={1}>
-          {columns.map((column, columnIndex) => (
-            <th key={columnIndex} aria-colindex={columnIndex + 1}>
-              {column.header}
-            </th>
-          ))}
-        </tr>
-      </thead>
-      <tbody>
-        {items.map((item, itemIndex) => (
-          <tr key={itemIndex} aria-rowindex={startIndex + itemIndex + 1 + 1}>
-            {columns.map((column, columnIndex) => (
-              <td key={columnIndex} aria-colindex={columnIndex + 1}>
-                {column.cell(item)}
-              </td>
-            ))}
-          </tr>
-        ))}
-      </tbody>
-    </table>
-  );
-}
-
-interface Item {
-  id: string;
-  name: string;
-  value: number;
-}
-
-const idColumn = { header: 'ID', cell: (item: Item) => item.id };
-const nameColumn = {
-  header: (
-    <span>
-      Name <button aria-label="Sort by name" />
-    </span>
-  ),
-  cell: (item: Item) => item.name,
-};
-const valueColumn = {
-  header: (
-    <span>
-      Value <button aria-label="Sort by value" />
-    </span>
-  ),
-  cell: (item: Item) => <ValueCell item={item} />,
-};
-const actionsColumn = {
-  header: 'Actions',
-  cell: (item: Item) => (
-    <span>
-      <button aria-label={`Delete item ${item.id}`} />
-      <button aria-label={`Copy item ${item.id}`} />
-    </span>
-  ),
-};
-
-const items: Item[] = [
-  { id: 'id1', name: 'First', value: 1 },
-  { id: 'id2', name: 'Second', value: 2 },
-  { id: 'id3', name: 'Third', value: 3 },
-  { id: 'id4', name: 'Fourth', value: 4 },
-];
-
-function ValueCell({ item }: { item: Item }) {
-  const [active, setActive] = useState(false);
-
-  return !active ? (
-    <span>
-      {item.value ?? 0} <button aria-label={`Edit value ${item.value}`} onClick={() => setActive(true)} />
-    </span>
-  ) : (
-    <span role="dialog">
-      <input value={item.value} autoFocus={true} aria-label="Value input" tabIndex={0} />
-      <button aria-label="Save" onClick={() => setActive(false)} />
-      <button aria-label="Discard" onClick={() => setActive(false)} />
-    </span>
-  );
-}
+import { TestTable, actionsColumn, idColumn, items, nameColumn, valueColumn } from './stubs';
 
 function getActiveElement() {
   return [
@@ -132,11 +13,6 @@ function getActiveElement() {
     document.activeElement?.textContent || document.activeElement?.getAttribute('aria-label'),
   ];
 }
-
-test('not activated for "table" role', () => {
-  const { container } = render(<TestTable tableRole="table" columns={[idColumn, nameColumn]} items={items} />);
-  expect(container.querySelectorAll('td[tabIndex],th[tabIndex]')).toHaveLength(0);
-});
 
 test('updates interactive elements tab indices', () => {
   const { container } = render(<TestTable columns={[nameColumn, valueColumn]} items={items} />);
@@ -258,46 +134,6 @@ test('does not throw errors if table is null', () => {
   expect(() => render(<TestTable />)).not.toThrow();
 });
 
-test('ensures table always has a user-focusable element', () => {
-  const { container } = render(<TestTable columns={[idColumn, valueColumn]} items={items} />);
-  const table = container.querySelector('table')!;
-  const row1 = table.querySelector('tr[aria-rowindex="2"]')!;
-  const sortButton = table.querySelector('button[aria-label="Sort by value"]') as HTMLElement;
-  const editButton = table.querySelector('button[aria-label="Edit value 1"]') as HTMLElement;
-
-  expect(table.querySelectorAll('[tabIndex="0"]')).toHaveLength(1);
-
-  sortButton.focus();
-  expect(table.querySelectorAll('[tabIndex="0"]')).toHaveLength(1);
-
-  sortButton.remove();
-  mockObserver.callback([
-    { type: 'childList', addedNodes: [], removedNodes: [sortButton] } as unknown as MutationRecord,
-  ]);
-  expect(table.querySelectorAll('[tabIndex="0"]')).toHaveLength(1);
-
-  editButton.focus();
-  expect(table.querySelectorAll('[tabIndex="0"]')).toHaveLength(1);
-
-  row1.remove();
-  mockObserver.callback([{ type: 'childList', addedNodes: [], removedNodes: [row1] } as unknown as MutationRecord]);
-  expect(table.querySelectorAll('[tabIndex="0"]')).toHaveLength(1);
-});
-
-test('ensures new interactive table elements are muted', () => {
-  const { container, rerender } = render(<TestTable columns={[idColumn, valueColumn]} items={items.slice(0, 3)} />);
-
-  expect(container.querySelectorAll('[tabIndex="-999"]')).toHaveLength(4);
-  expect(container.querySelectorAll('[tabIndex="0"]')).toHaveLength(1);
-
-  rerender(<TestTable columns={[idColumn, valueColumn]} items={items} />);
-  const lastRow = container.querySelector('[aria-rowindex="5"]')!;
-  mockObserver.callback([{ type: 'childList', addedNodes: [lastRow], removedNodes: [] } as unknown as MutationRecord]);
-
-  expect(container.querySelectorAll('[tabIndex="-999"]')).toHaveLength(5);
-  expect(container.querySelectorAll('[tabIndex="0"]')).toHaveLength(1);
-});
-
 test('ignores keydown modifiers other than ctrl', () => {
   const { container } = render(<TestTable columns={[nameColumn, valueColumn]} items={items} />);
   const table = container.querySelector('table')!;
@@ -318,40 +154,6 @@ test('ignores keydown modifiers other than ctrl', () => {
   expect(getActiveElement()).toEqual(['button', 'Sort by name']);
 
   fireEvent.keyDown(table, { keyCode: KeyCode.end, ctrlKey: true });
-  expect(getActiveElement()).toEqual(['button', 'Edit value 4']);
-});
-
-test('cell or cell element is re-focused after the focus target got removed', () => {
-  const { container } = render(<TestTable columns={[nameColumn, valueColumn]} items={items} />);
-  const table = container.querySelector('table')!;
-  const row1 = table.querySelector('tr[aria-rowindex="2"]')!;
-  const row2 = table.querySelector('tr[aria-rowindex="3"]')!;
-  const row3 = table.querySelector('tr[aria-rowindex="4"]')!;
-
-  expect(document.body).toHaveFocus();
-
-  row1.remove();
-  mockObserver.callback([{ type: 'childList', addedNodes: [], removedNodes: [row1] } as unknown as MutationRecord]);
-
-  expect(document.body).toHaveFocus();
-
-  table.querySelector('button')!.focus();
-  fireEvent.keyDown(table, { keyCode: KeyCode.down });
-  expect(getActiveElement()).toEqual(['td', 'Second']);
-
-  row2.remove();
-  mockObserver.callback([{ type: 'childList', addedNodes: [], removedNodes: [row2] } as unknown as MutationRecord]);
-
-  expect(getActiveElement()).toEqual(['td', 'Third']);
-
-  table.querySelector('button')!.focus();
-  fireEvent.keyDown(table, { keyCode: KeyCode.down });
-  fireEvent.keyDown(table, { keyCode: KeyCode.right });
-  expect(getActiveElement()).toEqual(['button', 'Edit value 3']);
-
-  row3.remove();
-  mockObserver.callback([{ type: 'childList', addedNodes: [], removedNodes: [row3] } as unknown as MutationRecord]);
-
   expect(getActiveElement()).toEqual(['button', 'Edit value 4']);
 });
 
@@ -421,7 +223,7 @@ test('elements focus is restored if table changes role after being rendered as g
 
   expect(getTabIndices()).toEqual([0, -999, -999, -999, -999]);
 
-  rerender(<TestTable tableRole="table" columns={[idColumn, valueColumn]} items={items} />);
+  rerender(<TestTable keyboardNavigation={false} columns={[idColumn, valueColumn]} items={items} />);
 
   expect(getTabIndices()).toEqual([0, 0, 0, 0, 0]);
 });
